feat(database): resolve merge conflicts
